requirements.txt
matrix-nio[e2e]==0.24.0
simplematrixbotlib==2.12.3
aiofiles
